added support for stream_span
stream_span has the advantage that it encapsulate a piece of
memory instead of containing one. Because of that it does not
own the memory, so you have to be a bit more carefull with it.

Another advantage is that because it does not own the memory
we can 'extract' a stream_span from a stream_span. Which in
case of parsing dvb data is actually a really nice and safe
way of doing things.

And as a bonus I could remove the crc_section subclass horror.